UAN-102: Initial commit of UAS.
This is the initial commit of the UAS framework.  The framework is
is generated from a swagger.yaml file.

The README.md file documents how to deploy the UAS on a SMS server as
part of a Kubernetes cluster.

The following endpoints are supported in this version.

/v1/uas_access
- This endpoint produces html output where all the
management functions of the UAN may be performed from a web browser.
UANs can be created, deleted, and a user can retrieve status about
any UANs they are running.

/v1/uan
- This endpoint is used by curl commands to perform the create and delete
functions from the command line.

/v1/uan?username=<user>&imagename=<image>
- Using the POST method, this endpoint will create a UAN for the user
using the given image.  If the imagename is omitted, a default image
will be used.  NOTE: At the time of this commit, there is no default
image set.

/v1/uan?uan_list=<uan1>,<uan2>
- Using the DELETE method, this endpoint will delete the UANs listed.
A list of available UANs to delete can be obtained by the endpoint
below.

/v1/uans?username=<user>
- Using the GET method, this endpoint will list all UANs for the user.

Because swagger requires Python 3, we have to use kubernetes 7.0.0b1 due
to a reserved word conflict between kubernetes 6.x.x and Python 3.  When
kubernetes officially releases 7.0.0, we need to update the
requirements.txt to reflect the non-beta release.

In order to run kubernetes from within the cray-user-access-service,
you need to create a kubernetes configmap of the
/etc/kubernetes/admin.conf file from the kubernetes master where you
deploy the cray-user-access-service.  The instructions for this are:

kubectl create configmap --from-file=/etc/kubernetes/admin.conf kube-cfg

The swagger.yaml file is under the schema directory and a Jenkinsfile.docs
was created in compliance with the Pubs group best practices for swagger
documentation.

The Jenkinsfile is compliant for building uan-mgr in the DST pipeline.