UAN-147: Adding ssh key support.
This commit adds support for gathering the users public ssh key
from the UAS webpage and via curl.  The key is passed to the UAN
container as an environment variable.

The UAN is also now deployed as a deployment and service.

In this version, only the crayadm user is supported.