UAN-164: Change uan-mgr to user k8s secret for cluster config.
This commit changes the uan-mgr from using a configmap to using
a secret for communicating the k8s config information.

The commit also passes this secret to the uans so eproxy and other
tasks interacting with k8s will function.

The commit also adds a default volume mount of the /scratch filesystem.
If this filesystem doesn't exist, an empty directory will be created.