UAN-181: Catch missing ssh key on create uan request.
This commit makes sure an ssh key is provided when creating a uan.

A style sheet change was needed to correct the font color on the
home page cancel button.  The previous button implementation was
not setting the required field on the ssh key properly.

Tested on the gui and curl command lines.