UAN-223: Adopt the naming convention for uas
Change uan-mgr so it aligns with the naming with the standard
that matches SMS components. uan-mgr will become cray-uas-mgr
and the container name will be cray/cray-uas-mgr.

Also change the label uas-mgr uses to search for the id service.
That service will be labeled cray-uas-id.

This is needed before we integrate our components into the
sms-installer.