UAN-183: Make uas-mgr configurable.
This commit adds configurability to the cray-uas-mgr.

The uas instances will now have an external IP address as configured
in the group_vars/all/networks.yaml file under gateways.user_gw.

A list of allowable uas images is also enforced and the default image
is configurable in the cray-uas-mgr template.

A list of volume mounts is now configurable in teh cray-uas-mgr template.

The configuration template file is in the sms_installer repo under the
cray-uas role.