CASMUSER-520: Add UAI affinity
Restrict UAI pods to nodes with the label uas=True. This uses
the same affinity that uas-mgr, uas-rex, and uas-img define
in the K8s manifest files.

Tested on my Craystack deployment by updating the container
image and restarting the pod. The UAI started on the correct
node because the label uas=True existed for a node.