CASMUSER-18: Improve status reporting of UAI
This commit improves upon the status reported to users on their UAIs.

If the status of the pod is running, the "Ready" condition is then
checked.  It is possible for the container to be in a Running state
but not Ready.  The status is now reported as "Running: Ready" or
"Running: Not Ready" if in the Running state.