CASMUSER-683: Add ansible role for cray-uas-mgr
Add the ansible role for cray-uas-mgr and cray-uas-id.
This role was previously run from the sms-installer repo.

To support the Ansible Framework deployment of Shasta services,
The deployment code will now live in the uas-mgr repo along
with the service. A new spec file and Jenkinsfile was added
to build an RPM for SLES.

Tested on pepsi-sms that the installed rpm added the ansible
tasks to the /opt/cray/crayctl/ansible_framework directories.

Ran ansible-playbook on:
/opt/cray/crayctl/ansible_framework/main/cray-uas-mgr.yml
and verified that the tasks completed successfully.