CASMUSER-679: Add cray-uas-mgr-auth service
To provide a preview of a service requiring auth,
add the service cray-uas-mgr-auth that requires
https and JWT for authentication. A new ansible role
cray_uas_mgr_auth is added that minimally duplicates
the role cray_uas_mgr. The notable exception is that
the sidecar container, register-api-gateway, sets the
values REQUIRE_AUTHENTICATION and REQUIRE_HTTPS to
'true'.

Tested on the pepsi-sms where ansible framework was
available. The service cray-uas-mgr-auth came up and
required https and requests without a JWT came back
as unauthorized. The cray-uas-mgr service also started
and did not require authentication.