CASMUSER-772: Refactor env variables from UAN_* to UAS_* and provide UAS_NAME
This commit refactors the env variables in the UAI from UAN_ prefixes
to UAS_ prefixes.  It also adds a new env variable called UAS_NAME which
the entrypoint.sh script may use to gather the various env values it
needs at runtime.  This removes the need to hardcode things like the
UAI port number in both the uas-mgr and entrypoint.sh.

Also, an unused function (create_service) was removed.