CASMUSER-775: Make UAI ports configurable
This commit adds support for configuring which ports a UAI should
expose as container ports as a NodePort service and service ports
as a ClusterIP service as well as an external IP for the ClusterIP
service.

The NodePort and ClusterIP services for a UAI are now labeled as
"uas_svc=<username>-<uuid>".

Support for mounting configmaps and secrets to the UAI are also
added. These mounts are needed for slurm client support.

The UAI name is now shortened to a <user>-<uuid> where <uuid> is a 8 hex
character identifier.  This change allows a user to create multiple UAIs
with the same image.

Unit tests are updated to include changed and new functions in uas_cfg.py.

This commit moves the handling of the uas_mgr.yml to the cray-uas-mgr.
It also updates the configmap-cray-uas-mgr.ymal.j2 template
to support variables that are used for port config and slurm.