CASMUSER-918: Add MetalLB IP pool annotations.
This commit adds the ability to set IP pool annotations in the UAI
services.  This is needed to select the proper IP pool provided by
MetalLB.

In the cray-uas-mgr-cfgmap, the following two variables are used to
select the MetalLB IP pool to use:

- uas_ssh_lb_pool: Sets the pool for ssh connection to the UAI
- uas_svc_lb_pool: Sets the pool for UAI services to use

These values are only used if the service type is set to LoadBalancer.

It is hoped that the values in vars/uas_mgr.yml can be set to Ansible
facts values once CMS has MetalLB configured in the system.  Dummy values
are set for now.