CASMUSER-1033: Expand localization coverage
This commit adds the ability to reconfigure nearly all the values in the
uas-mgr config map using the localization playbook.

Values that can be configured are controlled by the following ansible
variables which should be defined, if needed, in the customer_var.yml
file found under the customer_runbooks directory.  This is the place for
customers to localize their settings.

gwn_ipv4: sets the UAS external IP value.  This is required.

delete_uas_svc_ports: is a list port numbers or port ranges that will
be deleted from UAI use.

add_uas_svc_ports: is a list of port numbers or port ranges that will
be added to UAI use.

uas_default_img: will set the UAI default image value.

delete_uas_imgs: is a list of images to delete from UAI use.

add_uas_imgs: is a list of images to add for UAI use.

delete_uas_volume_mounts: is a list of volume mount specs to remove from
UAI use.

add_uas_volume_mounts: is a list of volume mount specs to add for UAI
use.