CASMUSER-690: add readiness check for UAIs
The UAI pod is declared ready by Kubernetes before the UAI is ready to
accept ssh connections. Add a k8s readiness check that will ensure that
sshd is listening on the specified port before declaring the Pod
available.

As a part of this change, set a const for the default port to make
changing the default easier in the future and enhance test coverage.
Additionally, it became clear when doing this that we cannot support
port ranges for uas_ports since sshd does not support them, so disable
that ability.

Closes: CASMUSER-690