CASMUSER-137: match UAI timezone to host
Add an optional type field for host_path mounts in order to support
FileOrCreate option. Although this can be used for other mount types as
well. The base use case for this is fixing the timezone bug.

By mounting in /etc/localtime into the container we can match the UAI
timezone to that of the host. This is set in the customizations config
since some users may want this disabled and also as a good example for
users.

Closes: CASMUSER-137