CASMUSER-619: Resiliency work for uas-mgr
- Increase replicas to 2 for minimal yet resource efficient resiliency
- Allow uas-mgr to run on any node in the SMS cluster
- Add pod anti-affinity to help spread out the 2 replicas

Customers only want/need to isolate UAI instances, the UAS
service can run as part of the normal SMS cluster of services.

Pod anti-affinity will prefer, but not require, the k8s scheduler to
spread out the replicas within the uas-mgr service. Preference was
chosen over required to avoid pending pods if we lose nodes.

This explicitly ignores the cray-uas-id service because it needs to be
tied to the node that has the LDAP information and is being deprecated
in the near future.

Closes-Bug: CASMUSER-619