CASMUSER-1216: Implement macvlan for UAIs
This commit enables the use of macvlan interfaces in UAIs.

It removes the LoadBalancer service which was used for INGRESS with
MetalLB as macvlan takes care of both INGRESS and EGRESS on the node
management network.

The uas_svc_ports are also removed as macvlan eliminates the need to
enumerate which ports are available on the node management network
interface.

The cray-uas-mgr role now installs and applies the necessary definitions
for the macvlan CNI.

This commit adds tasks to setup the necessary mac0 bridge on
sms03-nmn in support of macvlan operation.