CASMUSER-1005: better error handling on delete
Clean up the delete path a bit in a few ways:

1) Throw an error if any of the deployments that we passed in are not
found.

2) Handle the case where the user passes in the list with spaces in it,
as long as its quoted (unquoted was already an error)

3) Throw an error if the user passes in quoted spaces as this will look
like one large deployment that's not found.

This addresses most of the concerns in 1005 although it is still
possible to make it fail in odd ways - at least now it will fail with a
proper error rather than look like its going to work.