CASMUSER-1220: Add UasAuth class
Add a class to help with Auth related functions in advance of full
keycloak/JWT support.

This class will take a JWT and query keycloak's /userinfo endoint
to get the neccessary information to construct an /etc/passwd
entry to insert into the UAI.

Added unit tests to cover the simple functions in UasAuth. A follow
on commit will add a unit test to mock out the requests call to
/userinfo.

This commit mounts the ca cert required to use https when communicating
to keycloak.