CASMUSER-1434: Query Keycloak userinfo if token is sent with API requests
If a JWT is sent with a UAS API request, query Keycloak's /userinfo
endpoint to gather information about the user. uas-mgr will then
use that user information to create and list UAIs if it available.

As a fallback, username is still required with API requests but it
is ignored if a Bearer token is present. Once Auth is required for
UAS, the username parameter will be removed.