CASMUSER-1477: Add test tool deploy.sh to help cray-uas-mgr dev
This script is for dev only. It is not packaged in the rpm or
docker image.

This script will build, test, tag and deploy to sms-1.craydev.com
if all steps are successful.

The script assumes that you have modified the deployment spec of
cray-uas-mgr to use dtr.dev.cray.com/<USER>/cray-uas-mgr:latest
as the image.

It also helps to scale the number of replicas to one and remove
the liveliness probe for a sane logging experience.