CASMUSER-1497: Add cray_uas_user role
Add new cray_uas_user ansible play will use the kcadm.sh
utility to manually add users to keycloak. By default,
the ansible role will add the user uastest so that UAS will
function when auth is required but no user federation is configured
in Keycloak.

The play will loop over the users in uas_users as defined in
cray_uas_user/defaults/main.yml and user 'kcadm.sh create user'
to put the users in the shasta realm. If the user already exists,
the play will continue with the next user.

Tested on craystack that the play added the user and uastest could
use the UAS API to create, list, and delete UAIs. While this user
is not exclusive to UAS, it is meant to be provided as a reference
user for initial or internal use. Tested that the play could be run
more than once and was idempotent. If the play can not find the
keycloak pod all tasks will be skipped.

This role will need to be added to cme_standard_installer,
cme_premium_installer, and shastavm before the uastest user may be
used.