CASMUSER-1593 Restart uas-mgr after config change
On a few recent installations, UAIs created for CT weren't getting the
latest configuration, so Slurm wasn't working in the UAI. I suspect this
is due to the delay before config map updates are seen in pods.

After changing the configuration with the cray_uas_mgr_localize role,
restart the cray-uas-mgr pods. This makes sure the pods see the latest
version of the configmap.

Tested by running the Slurm play on Slice and immediately checking the
UAS configuration with cray uas images list.