CASMUSER-1615: Add role to add users to compute nodes
In order to support sbatch, the user submitting a batch job
must also be present on the compute nodes. Add an ansible role
to add the users listed in the uas_users var.

This ansible play will be documented as neccessary until users
are able to be federated in keycloak and available on compute
nodes.