CASMUSER-1667: check for invalid volume names
k8s limits names of things (like volumes) to be <= 63 chars and
generally DNS-name like, except no capital letters allowed. As a part of
this change, fix some copy/paste issues in the create_volume unit tests.

Also when I made this change, I switched an exception for bad volume
types to an abort so it's user visible.