CASMUSER-1469: delete/list UAIs by host/user
Add filters for the list all/delete all API calls so that admins will be
able to do the following:

- list UAIs by username
- list UAIs by host
- delete UAIs by username

This change also makes the list call more efficient. Since it uses the
label and/or field selectors, we have less processing to do on the UAS
manager side.

Finally we add a 40 second timeout with a k8s API call delay of 0.5
seconds when looking for the IP address for the UAI.

Closes-Bug: CASMUSER-1469
Closes-Bug: CASMUSER-1700