CASMUSER-1665: Scale out labeling of nodes and macvlan setup
This commit moves the labeling of nodes for UAI use to the
cray_uai_hosts role which is included in the UAS localization
role if the uas-mgr.yml playbook is used.  It can also be
run on its own directly.

UAI nodes were previously labeled during system install
and was static.  It used data in host_vars/<host> directories
which is going to be difficult to maintain when systems get larger.

Labeling of nodes is now based on hosts in a [uai_hosts] inventory
group.  This allows a single label and taints file to exist in
group_vars/uai_hosts for all UAI host nodes.

The labeling process is part of the cray_uai_hosts role which is
a localization type of role.  It can be run anytime after installation
to change the complement of nodes available for UAI hosting.

Along with labeling nodes for UAI use, this role will also
setup the macvlan bridge interface on each UAI host node.

This commit also adds support for the new networks.yml format and
the optional addition of a mountain cabinet node_management
network.