CASMUSER-1859: improve delete performance/status
Improve the delete process and pod status.

Complete deletes in the background, which improves the delete
performance especially for large batches.

This change is needed when deleting large amounts of UAIs at once, it
was taking on the order of minutes to finish the deletion.

Additionally, in cases where we create more pods than the system can
handle, many are left in a pending state. This change will ensure those
pods get a Pending status rather than no status like before.