CASMUSER-1886: Add motd and fix uan-interface gateway
This commit adds the ability to set the motd for UANs.

The motd file location may be set in customer_var.yml
as 'uan_motd_file' which holds the path to the motd file
to use.  If that variable is not defined in customer_var.yml,
a default motd file is used.

The commit also contains some gateway fixes for uan-interfaces.