CASMUSER-1982: Fix cray uas list failures
In some cases when a UAI fails to properly start "cray uas list" would
receive a 500 from the uas-mgr service. This is because the service
would try to set the IP in the uai_connection_string when we couldn't
get that info from Kubernetes.

This fix ensures that we no longer do that.