CASMUSER-2287: switch UAIs to a user namespace
This moves UAIs to a new namespace called user. This namespace is also
created here by the ansible code. Although this is configurable, moving
to a new namespace will effectively cause uas-mgr to lose track of
running UAIs and so should only be done before installation. It is left
configurable mainly as a defensive measure.