CASMUAS-37: Fix reference in helm template
A template for role and rolebindings was incorrectly referencing the
wrong path to get the namespace for UAIs. The end result was k8s
objects that should have been in the user namespace ended up in the
services namespace and calls to `cray uas list|create|delete` resulted
in Forbidden errors.