CASMPET-1491: Use Host from request headers to get userinfo
This change is required to create, list, and delete UAIs remotely
("from a laptop").

Previously, uas-mgr was hardcoded to use api-gw-service-nmn.local
to talk to keycloak when it needed to look up user information
like uid, gid, username, and home directory. Now that External DNS
is available, uas-mgr must use the same hostname that the token
was granted from (issuer or iss). It could be one of the following:
shasta.<system>.dev.cray.com
auth.<system>.dev.cray.com
api-gw-service-nmn.local

The issuer is now derived from the Host field in the request header.