CASMUAS-58: Add gen-uai-img.sh utility
Add a script to convert a compute image tar archive into a
docker image suitable for use with uas-mgr.

This script is only provided as a helper tool currently but
it will be used to transition away from building static
docker images in the pipeline.

Update connexion to fix a build failure.