CASMUAS-61: Add the cray_uai_img role
Add a role to generate a UAI image from the compute tar included
in the blob. This role is commented out in the runbook so it will
not run automatically yet.

This role is needed to generate a sles15sp1 UAI image. Previously
the UAI image could only be generated in the pipeline and required
many dependencies, took ~1 hour, and made it difficult to support
new sles versions of the compute image.

The new role takes approximately 5 minutes and changing sles version
can now be done by changing a single variable when the compute tar
is available.