PE-29504 Clean up macvlan reserved IPs on boot
If a host with macvlan set up is rebooted, the reserved IP addresses are
considered in-use by the macvlan CNI, because the reservation files in
/var/lib/cni/networks are left over from the previous boot. This means
that PBS and Slurm pods cannot start until those files are removed.

To work around this issue, create a tmpfiles.d file that will remove the
IP address reservation files at boot time. Tested by running updated
ansible play on surly2, rebooting w001, and verifying that the Slurm
pods started correctly.