CASMUAS-113: Remove cray_uai_hosts tasks and role from playbook
The tasks to configure a mac0 bridge route on each worker node
and creating the networkattachment definition are now done earlier
in the install or no longer needed. Remove them from this project.

Eventually cray_uai_hosts may be removed entirely but a dependency
with slurm and pbs needs to be dealt with first