CASMUAS-7 Temp update Dockerfile to use both DST and PyPi for pip
This is a temporary workaround until we can get the right packages
and versions into the DST pip index (DST-5496) so that the docker
build can work and pull in both the needed packages from PyPi and
the etcd3_model package from DST.  Before this is merged to master
we want to remove the PyPi override from the Dockerfile so that
only DST is used.

THe .gitignore and test-requirements.txt changes are fine to go
forward (noting that since this commit description may be on those
files going forward).