CASMUAS-7 Remove 'noqa' comments in code
The code now lints 100% cleanly and has targeted pylint comments
to address lint confusion, so there is no need for the more
cryptic 'noqa' comments, which were cluttering up the code and
don't seem to have any effect on pylint anyway.  This commit
removes these directives in favor of the clearer and more
effective pylint directives.