CASMUAS-175 switch from unsquash to mount for creating UAI images
The ansible for creating UAI images was using unsquashfs to make
them, however, unsquashfs does not preserve extended attributes on
files when unpacking the file system, so file capabilities (and any
other extended attributes) were being lost.  This change switches
to using a loopback mount to get access to the content of the
squashfs image instead of squeashfs.

This change also cleans up some shell script style that needed
fixing.