CASMUAS-210 Update K8s chart to match new macvlan subnet approach
The way macvlan subets are done in Shasta has changed with
CASMINST-1290 from having one supernet of compute nodes to having,
potentially, multiple subnets of compute nodes reached through the
gateway on the NMN of the host node.  This requires a change to
the UAS Helm chart macvlan template to support multiple routes
instead of one route that varies by system type.