CASMUSER-2685 Remove UAI image configuration from helm chart
Overview and Summary

The UAI image configuration in the helm chart is, by its nature,
always going to be wrong, or "latest" which is no longer avaialble.
Work has been done on the packages that provide UAI images to
register those images correctly, so the factory configuration no
longer contains this image information.  This changes the helm
charts, removes code that loads image registrations from the
configuration, and updates tests to no longer expect the factory
configuration to provide any image registrations.

Related Issues:

Partly completes: CASMUSER-2685

Testing:

Fill this out in the PR

Risks:

The risk for this is low, as the removed capability contributed
to incorrect UAI images being registered and set as default.